+++ lib/networking/api_provider.dart
@@ -18,6 +18,7 @@ import 'package:device_meta/device_meta.dart';
 import 'package:dio/dio.dart';
 import 'dart:io' show Platform;
 import 'package:flutter/foundation.dart' show kIsWeb;
+import 'package:nylo_framework/nylo_framework.dart';
 
 class ApiProvider {
   late Dio _dio;
@@ -157,7 +158,8 @@ class ApiProvider {
   /// HTTP GET request using a [url]
   Future<dynamic> get(url, {dynamic data}) async {
     try {
-      Response response = await _dio.get(url, queryParameters: data);
+      Response response =
+          await _dio.get(url, queryParameters: data, data: data);
       return response.data;
     } catch (error, stacktrace) {
       _printLog("$error stackTrace: $stacktrace");